updated release system

+++ scripts/release.sh
@@ -4,14 +4,21 @@ BUILD_DIR=$1
 RELEASE_DIR=$2
 NAME=$3
 
+if [ ! -f "$BUILD_DIR/$NAME.img" ]; then
+    echo "No current build to release, please run: make build"
+    exit -1
+fi
+
 echo "Copy image"
 cp "$BUILD_DIR/$NAME.img" "$RELEASE_DIR"
 
+cd $RELEASE_DIR
+
 echo "Making md5sum..."
-md5sum "$RELEASE_DIR/$NAME.img" > "$RELEASE_DIR/$NAME.img.md5"
+md5sum "$NAME.img" > "$NAME.img.md5"
 
 echo "Zipping..."
-zip "$RELEASE_DIR/$NAME.zip" "$RELEASE_DIR/$NAME.img" "$RELEASE_DIR/$NAME.img.md5"
+zip "$NAME.zip" "$NAME.img" "$NAME.img.md5"
 
 echo "Cleaning"
-rm "$RELEASE_DIR/$NAME.img" "$RELEASE_DIR/$NAME.img.md5"
+rm "$NAME.img" "$NAME.img.md5"
